Fixed some bugs and did some refactoring
I implemented some generic CBVs and added a signal for User Profile Creation. There was a bug that you can create superuser account without having a UserProfile. This is fixed now

I forgot to add password change button in the navbar and now it's added.

Also refactored some of the code to be more manageable.