Switchable dark theme and custom Http404
This update features something really cool! I added dark mode to the site. It is very easy to use it. All you need is a profile. Then just click Change theme and help your eyes! :)

Also this update features custom Http404 message, but IT is only shown when debug is False.
Also as side note I did some refactoring and file renaming to make it clearer what is what.