Some testing for accounts app and min length validator
In this update I tested some of the functionality of the app ```accounts```. It's not fully tested, but it's getting there.

Also I added a min length validator for username field in LoginForm.

Also as a last note I cleaned some HTML code. Now login.html is using modelforms_generic.html :)