Caching and DB indexing
This seems like a small update, but it's big.

First, I indexed the database with the things that I search. This can improve the performance of the database.

Secondly, and this is the big thing. I added caching for dark_mode context processor. Now every registered user is stored in the memory for 5 minutes. I did this, because everytime the user made a request, I needed to look in the DB.

BEWARE! There could be problems with the caching system, I tested it for several hours, but it's possible to have problems in production.