First batch of
This update is really huge! It is not fully done yet, so expect a lot of refactoring, front end and back end improvements!

In this update, I implemented a Team system. This mean that you can create your team and invite your friends! This update gives me room for soooo much expansion, that I can't express it. Expect in the future really cool new features with the teams!