Team Todos, optimization, refactoring
This is a very interesting update! In this update I added team todos. If you are a member of a team, you can create a todo for this team!

Also, for optimization purpose, I made a simple middleware for caching. A lot has changed, and a lot will be changed!

Note: This is a release for the staging branch. DO NOT USE in production!