HTML Refactoring and redirecting
This is relatively small update, but includes a lot of refactoring!

Now when searching in /accounts URL, it's not using query strings to show the result. Instead it is redirecting! This design is much cleaner than using query strings.

Also I added a feature to check user profiles from `teams` section. It is an anchor tag leading to /accounts/USER.